feat(explorer): move getsecondsfrominterval to library
