+++ apps/explorer/src/app/components/epoch-overview/epoch-missing.spec.tsx
@@ -1,10 +1,7 @@
 import { MockedProvider } from '@apollo/client/testing';
 import { render } from '@testing-library/react';
-import EpochMissingOverview, {
-  calculateEpochData,
-  getSecondsFromInterval,
-} from './epoch-missing';
-
+import EpochMissingOverview, { calculateEpochData } from './epoch-missing';
+import { getSecondsFromInterval } from '@vegaprotocol/react-helpers';
 const START_DATE_PAST = 'Monday, 17 February 2022 11:44:09';
 
 describe('getSecondsFromInterval', () => {
